requirements.txt
-e .
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_elasticloadbalancingv2
aws-cdk.aws_batch
boto3
